+++ drivers/wifi/nrf71/CMakeLists.txt
@@ -9,37 +9,27 @@ function(nrf71_wifi_setup_include_and_sources)
 
   set(nrf71_osal_base ${CMAKE_CURRENT_SOURCE_DIR}/osal)
   set(nrf71_os_shim_dir ${CMAKE_CURRENT_SOURCE_DIR}/os)
-  set(nrf71_bus_dir ${CMAKE_CURRENT_SOURCE_DIR}/bus)
   set(os_agnostic_base ${ZEPHYR_NRF_WIFI_MODULE_DIR})
   set(fw_bins_base ${ZEPHYR_NRF_WIFI_MODULE_DIR}/zephyr/blobs/wifi_fw_bins)
 
   zephyr_library_include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}
     ${CMAKE_CURRENT_SOURCE_DIR}/inc
-    ${CMAKE_CURRENT_SOURCE_DIR}/../bus
     ${CMAKE_CURRENT_SOURCE_DIR}/inc/common/fw_if
     ${nrf71_osal_base}/os_if/inc
-    ${nrf71_osal_base}/bus_if/bus/qspi/inc
-    ${nrf71_osal_base}/bus_if/bal/inc
     ${nrf71_osal_base}/fw_if/umac_if/inc
     ${nrf71_osal_base}/hw_if/hal/inc
     ${nrf71_os_shim_dir}
-    ${nrf71_bus_dir}
     ${ZEPHYR_BASE}/subsys/net/ip
   )
 
-  # ---- Sources: osal (bus_if, fw_if, hw_if) ----
+  # ---- Sources: osal (fw_if, hw_if) and common ----
   zephyr_library_sources(
     ${nrf71_osal_base}/os_if/src/osal.c
     ${nrf71_osal_base}/hw_if/hal/src/common/hal_api_common.c
-    ${nrf71_osal_base}/bus_if/bal/src/bal.c
-    ${nrf71_osal_base}/bus_if/bus/qspi/src/qspi.c
     ${nrf71_osal_base}/fw_if/umac_if/src/common/fmac_cmd_common.c
     ${nrf71_osal_base}/fw_if/umac_if/src/common/fmac_api_common.c
     ${nrf71_os_shim_dir}/shim.c
-    ${nrf71_bus_dir}/device.c
-    ${nrf71_bus_dir}/ipc_if.c
-    ${nrf71_bus_dir}/ipc_service.c
     src/common/rf_params.c
     src/common/util.c
     src/common/vtf.c
@@ -48,6 +38,7 @@ function(nrf71_wifi_setup_include_and_sources)
     src/common/work_mgmt.c
     src/common/lock_mgmt.c
     src/common/llist_mgmt.c
+    src/common/ipc_bus.c
   )
 
   if(CONFIG_NRF71_SYSTEM_MODE OR CONFIG_NRF71_SCAN_ONLY)
@@ -76,9 +67,7 @@ function(nrf71_wifi_setup_include_and_sources)
       ${CMAKE_CURRENT_SOURCE_DIR}/inc/common/fw_if
       ${CMAKE_CURRENT_SOURCE_DIR}/inc
       ${nrf71_osal_base}/fw_if/umac_if/inc
-      ${nrf71_osal_base}/os_if/inc
-      ${nrf71_osal_base}/bus_if/bal/inc
-      ${nrf71_osal_base}/bus_if/bus/qspi/inc
+      ${nrf71_osal_base}/hw_if/hal/inc
     )
   endif()
