+++ drivers/wifi/nrf71/osal/hw_if/hal/inc/offload_raw_tx/hal_api.h
@@ -13,7 +13,6 @@
 #ifndef __HAL_API_OFF_RAW_TX_H__
 #define __HAL_API_OFF_RAW_TX_H__
 
-#include <bal_api.h>
 #include <common/hal_structs_common.h>
 #include <common/hal_api_common.h>
