+++ drivers/wifi/nrf71/osal/hw_if/hal/inc/radio_test/hal_api.h
@@ -15,7 +15,6 @@
 #define __HAL_API_RT__H__
 
 #include <nrf71_wifi_ctrl.h>
-#include "bal_api.h"
 #include "common/hal_structs_common.h"
 #include "common/hal_api_common.h"
