+++ drivers/wifi/nrf71/src/offload_raw_tx/api.c
@@ -20,6 +20,7 @@
 #include <common/fw_if/nrf71_wifi_rf.h>
 #include <common/rf_params.h>
 #include <common/util.h>
+#include <osal_api.h>
 #include <offload_raw_tx/api.h>
 
 #define DT_DRV_COMPAT nordic_wlan
